Add files via upload
 Robert Gholson 12/09/2024
My Data Set
Conditions Contributing to COVID-19 Deaths, by State and Age, Provisional 2020-
2023
As I was collecting this data, I found information drawn from a csv on Covid-19 and the
age groups that were affected the most. I found that states that had a good number of
documented mentions of Covid -19 through each state from years 2020-2023. For this
dataset I developed 6 sheets covering data from different visualizations and they show
data for age groups, diseases that lead to covid deaths, and the number of mentions of
covid symptoms. I was able to Utilize a map feature to show the different mentions in
each state for suspected covid cases. There are different charts, one is caused by
induce illnesses and affected deaths is caused by different and another with increased
and decreased of deaths through the years. I have a bar chart that also shows the
effected numbers of people that has died in this dataset. There are highlights
throughout the maps Of Number of people effected in different States. I hope to show
where Covid affected people the most in the United States and what was there cause of
death induced by with covid 19. I found this dataset interesting because this is not an
epidemic, this was a pandemic, and it wiped out a lot of people and messed up a lot of
people up. Something this major should be studied and researched so I figured I’d
utilized this dataset to understand more about how to do the research on this topic. This
information is important because you need to know this data too maybe be ready for the
next wave of a pandemic or how to have the proper medicine of restrictions in place for
people that can be affected in these situations. My hypotheses are that covid went over
10,000,000 reported cases in the U.S alone.